pyproject.toml
[tool.poetry]
name = "biomed"
version = "0.1.0"
description = "A bot for analyzing"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
pydantic = "^2.9.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
ruff = "^0.7.2"
asyncpg = "^0.30.0"
boto3 = "^1.35.63"
aiogram = {version = "^3.14.0", python = ">=3.9,<4.0"}
apscheduler = "^3.11.0"
pytz = "^2024.2"


[tool.poetry.group.dev.dependencies]
alembic = "^1.13.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
